backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pytest
httpx
langgraph
langchain
langchain-community
langchain-google-genai
langsmith
duckduckgo-search